implement mopage trigger behavior.
The trigger behavior for the event folder provides configuration
fields for setting up automatic notification to the mopage API whenever
an event page is published with ftw.publisher (receiver side).

This requires an ftw.publisher setup as well as a configured
collective.taskqueue.
The notification will be triggered when the ftw.publisher.receiver has
received a event update.