Show inactive events if the user has the permission to add events.
Inactive event pages will now be shown in the event listing view (and thus the event listing block) if the user has the permission to add event pages..

Important: the event pages will only be shown if event pages from a single container are shown. Otherwise a user could trick the system and be able to see inactive event pages from a path where he does not have the permission to add events.