Fix event archive portlet compatibility with ftw.news
The functionality of the event archive portlet was copied from the news archive portlet. The html classes in both (ftw.news and ftw.events) were the same which meant that when ftw.news was installed next to ftw.events, both javascripts acted on both archive portlets. This had the effect that on click, the collapsible months were opened and immediately closed again.

By adding unique classes the news archive portlet javascript will not act on the event archive portlet anymore.